Project-1
When your solar panels collect solar energy, the process produces a higher output than your batteries can handle. For your system to work, you need to control the flow into the battery to get the most efficient flow and storage possible. A charge controller accomplishes this for your system. The MPPT solar charge controller is a DC-to-DC converter for your solar power system. It receives voltage from the solar panels and converts it to charge your battery at a more appropriate level.